Article controller tests implemented

+++ backend/src/main/kotlin/com/jetbrains/life_science/article/master/service/ArticleServiceImpl.kt
@@ -3,8 +3,11 @@ package com.jetbrains.life_science.article.master.service
 import com.jetbrains.life_science.article.master.entity.Article
 import com.jetbrains.life_science.article.master.factory.ArticleFactory
 import com.jetbrains.life_science.article.master.repository.ArticleRepository
+import com.jetbrains.life_science.article.version.service.ArticleVersionService
 import com.jetbrains.life_science.category.service.CategoryService
+import com.jetbrains.life_science.exception.ArticleNotEmptyException
 import com.jetbrains.life_science.exception.ArticleNotFoundException
+import org.springframework.beans.factory.annotation.Autowired
 import org.springframework.stereotype.Service
 import org.springframework.transaction.annotation.Transactional
 
@@ -14,6 +17,9 @@ class ArticleServiceImpl(
     val categoryService: CategoryService,
     val repository: ArticleRepository
 ) : ArticleService {
+
+    @Autowired
+    lateinit var versionService: ArticleVersionService
 
     @Transactional
     override fun create(info: ArticleInfo): Article {
@@ -40,6 +46,9 @@ class ArticleServiceImpl(
     }
 
     override fun deleteById(articleId: Long) {
+        if (versionService.getByArticleId(articleId).isNotEmpty()) {
+            throw ArticleNotEmptyException("Article with id $articleId is not empty")
+        }
         existById(articleId)
         repository.deleteById(articleId)
     }
